connexion server / client wip

+++ includes/utils.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.hpp                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tbrebion <tbrebion@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rvrignon <rvrignon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/26 23:17:29 by flcarval          #+#    #+#             */
-/*   Updated: 2023/02/27 12:03:42 by tbrebion         ###   ########.fr       */
+/*   Updated: 2023/03/09 17:03:29 by rvrignon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include <stdlib.h>
 # include <csignal>
 
-long	parsing_args(int ac, char **av, char **env);
+int		parsing_args(int ac, char **av, char **env);
 void	sig_handler(int sig);
 
 #endif
